code format and package organized along with proper comments.
